created a hook for reading story files. created a story book component to hold ink logic.
