+++ src/StoneTower/hooks/useFetchStoryJSON.ts
@@ -1,0 +1,13 @@
+import { useEffect, useState } from 'react';
+
+export const useFetchStoryJSON = (fileName: string) => {
+    const [rawStory, setRawStory] = useState<null | any>(null);
+
+    useEffect(() => {
+        import(`../../assets/${fileName}.ink.json`).then((rawStory) =>
+            setRawStory(rawStory.default),
+        );
+    }, []);
+
+    return rawStory;
+};
